+++ apps/opplaringspenger-soknad/src/app/utils/appEnv.ts
@@ -4,7 +4,7 @@ import { AppEnv } from '../../../env.schema';
 export const getAppEnv = (): AppEnv => ({
     ...getCommonEnv(),
     ...getK9SakInnsynEnv(),
-    SIF_PUBLIC_FEATURE_NYNORSK: getMaybeEnv('SIF_PUBLIC_FEATURE_NYNORSK') === 'on' ? 'on' : 'off',
+    INJECT_DECORATOR: getMaybeEnv('INJECT_DECORATOR') === 'true' ? 'true' : 'false',
 });
 
 export const appEnv = getAppEnv();