+++ apps/opplaringspenger-soknad/src/msw/mocks/soker1/søkerMock.ts
@@ -1,8 +1,0 @@
-export const søkerMock = {
-    aktørId: '2320509955297',
-    fødselsdato: '1995-06-02',
-    fødselsnummer: '02869599258',
-    fornavn: 'Test',
-    mellomnavn: null,
-    etternavn: 'Brukeresen',
-};