+++ apps/pleiepenger-sykt-barn/src/app/søknad/opplysninger-om-barnet-step/RegistrertBarnPart.tsx
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 import { useAppIntl } from '@i18n/index';
 import { RegistrertBarn } from '@navikt/sif-common-api';
 import { resetFieldValues, SkjemagruppeQuestion } from '@navikt/sif-common-formik-ds';
-import { AnnetBarnValue, VelgBarnInput } from '@navikt/sif-common-ui';
+import { VelgBarn_AnnetBarnValue, VelgBarnFormPart } from '@navikt/sif-common-forms-ds';
 import { getRequiredFieldValidator } from '@navikt/sif-validation';
 import { useFormikContext } from 'formik';
 import { initialValues, SøknadFormField, SøknadFormValues } from '../../types/søknad-form-values/SøknadFormValues';
@@ -19,10 +19,10 @@ const RegistrertBarnPart = ({ søkersBarn }: Props) => {
         setFieldValue,
     } = useFormikContext<SøknadFormValues>();
 
-    const søknadenGjelderEtAnnetBarn = barnetSøknadenGjelder === AnnetBarnValue;
+    const søknadenGjelderEtAnnetBarn = barnetSøknadenGjelder === VelgBarn_AnnetBarnValue;
 
     useEffect(() => {
-        if (barnetSøknadenGjelder !== AnnetBarnValue) {
+        if (barnetSøknadenGjelder !== VelgBarn_AnnetBarnValue) {
             resetFieldValues(
                 [
                     SøknadFormField.barnetsFødselsnummer,
@@ -40,7 +40,7 @@ const RegistrertBarnPart = ({ søkersBarn }: Props) => {
 
     return (
         <SkjemagruppeQuestion legend="Barn" hideLegend={true}>
-            <VelgBarnInput
+            <VelgBarnFormPart
                 name={SøknadFormField.barnetSøknadenGjelder}
                 legend={text('steg.omBarnet.hvilketBarn.spm')}
                 registrerteBarn={søkersBarn}