+++ apps/pleiepenger-sykt-barn/src/app/utils/formValuesToSøknadsdata/__tests__/extractBarnSøknadsdata.test.ts
@@ -1,7 +1,7 @@
-import { extractBarnSøknadsdata } from '../extractBarnSøknadsdata';
-import { OmBarnetFormValues } from '../../../types/søknad-form-values/SøknadFormValues';
+import { VelgBarn_AnnetBarnValue } from '@navikt/sif-common-forms-ds';
 import { BarnRelasjon, ÅrsakManglerIdentitetsnummer } from '../../../types';
-import { AnnetBarnValue } from '@navikt/sif-common-ui';
+import { OmBarnetFormValues } from '../../../types/søknad-form-values/SøknadFormValues';
+import { extractBarnSøknadsdata } from '../extractBarnSøknadsdata';
 
 const formValues: OmBarnetFormValues = {
     barnetsNavn: '',
@@ -24,7 +24,7 @@ describe('extractBarnetSøknadsdata', () => {
         it('returnerer annetBarn dersom bruker la til barn med fnr', () => {
             const result = extractBarnSøknadsdata({
                 ...formValues,
-                barnetSøknadenGjelder: AnnetBarnValue,
+                barnetSøknadenGjelder: VelgBarn_AnnetBarnValue,
                 barnetsNavn: 'Test Testen',
                 barnetsFødselsnummer: '12345678911',
                 relasjonTilBarnet: BarnRelasjon.FAR,
@@ -38,7 +38,7 @@ describe('extractBarnetSøknadsdata', () => {
         it('returnerer annetBarnUtenFnr dersom bruker la til barn uten fnr', () => {
             const result = extractBarnSøknadsdata({
                 ...formValues,
-                barnetSøknadenGjelder: AnnetBarnValue,
+                barnetSøknadenGjelder: VelgBarn_AnnetBarnValue,
                 barnetsNavn: 'Test Testen',
                 barnetHarIkkeFnr: true,
                 barnetsFødselsnummer: '',
@@ -56,7 +56,7 @@ describe('extractBarnetSøknadsdata', () => {
         it('returnerer undefined med tomt barnet', () => {
             const result = extractBarnSøknadsdata({
                 ...formValues,
-                barnetSøknadenGjelder: AnnetBarnValue,
+                barnetSøknadenGjelder: VelgBarn_AnnetBarnValue,
                 barnetsNavn: 'Test Testen',
                 barnetHarIkkeFnr: true,
                 barnetsFødselsnummer: '',