+++ apps/pleiepenger-sykt-barn/src/app/utils/formValuesToSøknadsdata/extractBarnSøknadsdata.ts
@@ -1,17 +1,17 @@
-import { AnnetBarnValue } from '@navikt/sif-common-ui';
+import { VelgBarn_AnnetBarnValue } from '@navikt/sif-common-forms-ds';
 import { BarnRelasjon, ÅrsakManglerIdentitetsnummer } from '../../types';
 import { OmBarnetFormValues } from '../../types/søknad-form-values/SøknadFormValues';
 import { OmBarnetSøknadsdata } from '../../types/søknadsdata/OmBarnetSøknadsdata';
 
 export const extractBarnSøknadsdata = (values: OmBarnetFormValues): OmBarnetSøknadsdata | undefined => {
-    if (values.barnetSøknadenGjelder && values.barnetSøknadenGjelder !== AnnetBarnValue) {
+    if (values.barnetSøknadenGjelder && values.barnetSøknadenGjelder !== VelgBarn_AnnetBarnValue) {
         return {
             type: 'registrerteBarn',
             aktørId: values.barnetSøknadenGjelder,
         };
     }
 
-    if (values.barnetSøknadenGjelder === AnnetBarnValue) {
+    if (values.barnetSøknadenGjelder === VelgBarn_AnnetBarnValue) {
         if (values.barnetsFødselsnummer) {
             return {
                 type: 'annetBarn',