+++ packages/sif-common-forms-ds/src/form-parts/velg-barn-form-part/i18n/nn.ts
@@ -1,0 +1,8 @@
+export const velgBarnFormPart_nn = {
+    '@forms.velgBarnFormPart.legend': 'Barn registrert på deg',
+    '@forms.velgBarnFormPart.gjelderAnnetBarn': 'Søknaden gjelder et annet barn',
+    '@forms.velgBarnFormPart.født': 'Født {dato}',
+    '@forms.velgBarnFormPart.annetBarn.tittel': 'Annet barn',
+    '@forms.velgBarnFormPart.kilde':
+        'Lista viser barn som er registrert på deg i Folkeregisteret. Viss dette ikkje stemmer, kan du kontakte Skatteetaten.',
+};