+++ packages/eslint-config-sif/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sif/eslint-config
 
+## 1.1.37
+
+### Patch Changes
+
+- Pakkeoppdateringer
+
+## 1.1.36
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/major + zod4
+
 ## 1.1.35
 
 ### Patch Changes